Make use of storage class for new decl-specifier instead of string magic.
mutable storage class detection doesn't work, but typedefs seem to work
now. Detection must be improved, but it should be investigated if a
tree-based refactoring is somehow feasible, even so the string-based
approach seems to be working (most of the time).

I believe copying the decl-specifier with the declarator capturing the
name and removing the  array-declarator could actually work to retrieve
the correct array member type. copying the expression node with position
information could also help