Implement custom RefactoringContext
The standard CRefactoringContext shipped with CDT takes all open
projects in a workspace into account for symbol resolution. This causes
issues when, for example, two independent projects define the same
struct where one of the structs does implement a given member function
while the other doesn't. Depending on the order in which the index was
built, CDT might see the struct with the member function first, and
thus the refactoring to add a non existing member function fails
dramatically.